refactor(lite): 重构多个组件并提升代码质量
-优化 Bitmap 扩展函数，增加图片压缩质量参数
- 改进旋转图片函数，优化内存使用
- 更新适配器代码，提高可读性和性能
-修复 HtmlRenderEngine 中的图片处理逻辑
- 优化屏幕适配和 Glide 使用

+++ app/src/main/java/com/pengxh/kt/lib/fragments/extensions/BitmapExtensionFragment.kt
@@ -64,7 +64,7 @@ class BitmapExtensionFragment : KotlinBaseFragment<FragmentExtensionBitmapBindin
             if (binding.angleView.text.isNullOrBlank()) {
                 return@setOnClickListener
             }
-            val rotateImage = originalBitmap.rotateImage(binding.angleView.text.toString().toInt())
+            val rotateImage = originalBitmap.rotateImage(binding.angleView.text.toString().toFloat())
             binding.revolvedImageView.setImageBitmap(rotateImage)
         }
